Allow rendering the suggestion results with HTML
This is compatible with rendering text but also enables the `queryFn`
to highlight the result text in the result.

For example:

```js
  item.text = item.text.replace(new RegExp('(' + query + ')', 'ig'), '<b>$1</b>');
```

Also, since the property to render is controlled by the `queryFn`,
and the default `queryFn` returns hardcoded 'text', use the same
property here.
This fixes the bug of using a custom `textProperty` which now
renders empty results.